Made a basic working prototype
Have made a working prototype without much polish the ui is still not good will probably make improvements later. Also another issue that i am going to solve later is the name of the fruits as they are currently only labeled. The last part is related to the use of multiple voice lines which i will figure out later